% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addWatermark}
\alias{addWatermark}
\title{Add water mark}
\usage{
addWatermark(txt = NULL, cex = par("cex.axis") * 0.5)
}
\arguments{
\item{txt}{character string to be displayed}

\item{cex}{character expansion factor}
}
\description{
Function to plot a watermark stamp to the bottom right of the plot
}
\examples{
plot(rnorm(100))
addWatermark()
}
\keyword{aplot}

