% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allWorld}
\alias{allWorld}
\alias{allWorldKey}
\title{Giorgi regions overview plots}
\usage{
allWorld(fun, regnames = NULL, inset.bg = "#FFFFFFCC", ...)

allWorldKey(fun, ...)
}
\arguments{
\item{fun}{function to be executed at the respective spots for insets}

\item{regnames}{which regions should be plotted (see details)}

\item{inset.bg}{background colour for insets (with transparency)}

\item{...}{additional arguments passed to \code{fun}}
}
\description{
This function provides the backbone for the global overview plots.
In essence, it provides the map and the locations where to add insets.
Also \code{allWorldKey} offers the additional functionality of
placing a legend in the bottomleft corner.
}
\details{
The list for background colouring has to contain at least the object
\code{data}, a vector with values to use in colouring named with the region
abbreviations. In addition, \code{col} and \code{lev} can be added to control
the colouring. \code{title} will add a title and \code{units}
will add units to the legend.
}
\keyword{plot}

