\docType{data}
\name{NRMcolours}
\alias{NRMcolours}
\alias{blackcol}
\alias{heatcols}
\alias{heatcols2}
\alias{modcol}
\alias{modcol2}
\alias{obscol}
\alias{oldcolours}
\alias{scencols}
\alias{scencols2}
\alias{scencols3}
\alias{scencolvals}
\alias{wetcols}
\alias{wetcols2}
\title{NRM colours}
\format{Named vectors with hex colour codes.}
\description{
A dataset containing the colours used in the NRM report.
This dataset includes: \describe{ \item{scencols}{Scenario
colours used for bars and inner shaded area}
\item{scencols2}{Scenario colours for median value
(darker)} \item{scencols3}{Scenario colours for outer
shaded area (lighter)} \item{modcol,modcol2}{Colour for
model overlays} \item{obscol}{Colour for observations}
\item{blackcol}{Colour for axes and text}
\item{heatcols,heatcols2}{Dark and light shades of colours
for positive and negative values (red-blue)}
\item{wetcols,wetcols2}{Dark and light shades for rainfall
related variables (blue-brown)} }
}
\keyword{datasets}

