% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{nrmexample}
\alias{nrmexample}
\title{NRM test data for plots}
\format{lists of NetCDF objects}
\description{
A dataset containing simulated and observed temperature time
series and temperature change percentiles for examples.
}
\keyword{datasets}

