% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotcolbar}
\alias{plotcolbar}
\title{Function to add a single coloured bars}
\usage{
plotcolbar(x, y, yshade = NULL, col = hcl(c(240, 0), l = 60, c = 50),
  col2 = hcl(c(240, 0), l = 30, c = 80), lwd = 10)
}
\arguments{
\item{x}{location on the horizontal axis to plot}

\item{y}{vector of percentiles (e.g. 10, 90, 50th)}

\item{yshade}{value for inner shading (see details)}

\item{col}{colour for negative and positive values}

\item{col2}{colour for horizontal line (median)}

\item{lwd}{line width of bar}
}
\description{
Plot bar of projected changes at specified location with light
shading of inner region (generally used to indicate natural variability).
}
\details{
Inner shading can be indicated both using a single value (per bar)
or two values per bar for bottom and top extend if asymmetrical.
}
\keyword{plot}

