\name{projbar}
\alias{projbar}
\title{Add multiple projections bars (with extensions)}
\usage{
projbar(x, y, y2 = NULL, col = "grey", col2 = "black", lwd = 10,
  lwd2 = 2)
}
\arguments{
  \item{x}{horizontal intersects for bars}

  \item{y}{matrix with array of bars (percentiles; bars)}

  \item{y2}{matrix with extensions to bars}

  \item{col,col2}{colours for bar shading and vertical line
  (median)}

  \item{lwd,lwd2}{bar width and extensions width}
}
\description{
Simple function to add multiple projections bars with
extensions to existing plot.
}
\keyword{plot}

